+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-## dbt-databricks 1.7.9 (TBD)
+## dbt-databricks 1.8.0 (TBD)
+
+### Features
+
+- Support `on_config_change` for materialized views, expand the supported config options ([536](https://github.com/databricks/dbt-databricks/pull/536)))
+- Support `on_config_change` for streaming tables, expand the supported config options ([569](https://github.com/databricks/dbt-databricks/pull/569)))
+
+## dbt-databricks 1.7.9 (Mar 5, 2024)
 
 ### Fixes